==> R/02_A_beta-countries.qmd <==
---
title: "02_D_beta-countries"
subtitle: "Analysis of mismatch between country native composition and primary type data"
format: html
eval: FALSE
---

# Reading packages

```{r packages}
# data
library(readr)        # reading CSV files
library(here)         # constructing file paths
library(dplyr)        # data manipulation
library(tidyr)        # data tidying
library(phyloregion)  # handling phylogenetic data and transformations
```

# Importing and processing data 

Importing and processing native composition and NBT composition data

```{r}
# From 01_C_data_preparation.qmd
spp_native_distribution <- readr::read_csv(here::here("data", "processed", "spp_native_distribution.csv")) 

# From 01_C_data_preparation.qmd
spp_type_distribution <- readr::read_csv(here::here("data", "processed", "spp_type_distribution.csv")) 
```

Checking the specimens with data in both tables (native distribution and types). We transformed the long format species occurrence data frame to dense format. During this procedure we removed 1204 species that do not have information on native distribution (or we couldn't get this information from CAS)

```{r}
df_native_grid <- 
  spp_native_distribution |> 
  dplyr::select(grids = country_distribution, 
                species = valid_name) |> 
  tidyr::drop_na(grids)

df_type_grid <- 
  spp_type_distribution |> 
  dplyr::select(grids = country_museum, 
                species = valid_name) |> 
  tidyr::drop_na(grids) |> 
  dplyr::mutate(grids = paste(grids, "type", sep = "_"))
  

# joining data frames
df_all_grid <- rbind(df_native_grid, df_type_grid) # joining both matrices - 
    #native and types composition

#### Just descriptive quantities
country_native <- unique(df_native_grid$grids)
country_type <- gsub(pattern = "_type", 
                     replacement = "",
                     unique(df_type_grid$grids))
country_type_zero <- setdiff(country_native, country_type) # countries with no type specimen

# transforming into a sparse matrix to speed up calculations
sparse_all <- df_all_grid |> 
  phyloregion::long2sparse(grids = "grids", species = "species") |> 
  phyloregion::sparse2dense()

# Transforming in presence absence matrix
sparse_all_pa <- ifelse(sparse_all >= 1, 1, 0) 

# Binding countries with no types - adding zeroes
country_type_zero_names <- paste(country_type_zero, "_type", sep = "") # this will be used to bind together matrix with types and add the countries with no type
matrix_type_zero <- matrix(0,
         nrow = length(country_type_zero_names),
         ncol = ncol(sparse_all_pa), 
         dimnames = list(country_type_zero_names, 
                         colnames(sparse_all_pa)))

sparse_all_pa2 <- rbind(sparse_all_pa, matrix_type_zero)

```

# Calculating directional turnover based on native and primary type comparison

Here we calculated the turnover in two directions. One is the turnover of native composition, in other words, the underrepresentation of native fish species in museums and natural collections within the country. Values closer to one indicate that the country present a huge underepresentation of its native fish fauna in primary types located within the country. 

The other metric is primary type turnover. Values approaching one indicate that there is an overepresentation of primary types maintained in the country when compared to the native fish fauna of that country.

```{r}
source(here::here("R", "function_beta_types.R"))

names_countries <- unique(df_native_grid$grids) # country names

df_all_beta <- beta_types(presab = sparse_all_pa2, 
                          names.countries = names_countries) # calculating metrics of directional turnover

readr::write_csv(df_all_beta, here::here("data", "processed", "df_all_beta.csv"))
```
